feat: adds testingSelectors export
