+++ packages/odyssey-react-mui/package.json
@@ -18,6 +18,9 @@
       ],
       "labs": [
         "./dist/src/labs/index.d.ts"
+      ],
+      "testingSelectors": [
+        "./dist/src/testingSelectors/index.d.ts"
       ]
     }
   },
@@ -30,6 +33,10 @@
       "types": "./dist/src/labs/index.d.ts",
       "default": "./dist/labs/index.js"
     },
+    "./testingSelectors": {
+      "types": "./dist/src/testingSelectors/index.d.ts",
+      "default": "./dist/testingSelectors/index.js"
+    },
     ".": {
       "types": "./dist/src/index.d.ts",
       "default": "./dist/index.js"
@@ -41,8 +48,9 @@
     "build:cleanIconComponents": "rimraf src/icons.generated",
     "build:cleanIconsIndex": "rimraf src/icons.generated/index.ts",
     "build:source": "NODE_ENV=production BROWSERSLIST_ENV=modern babel src --out-dir dist --extensions .js,.jsx,.ts,.tsx --source-maps",
+    "build:testingSelectors": "tsx scripts/generateTestingSelectorsJson.ts",
     "build:types": "tsc --project tsconfig.production.json",
-    "build": "npm-run-all --sequential build:clean generate:i18n build:types build:source",
+    "build": "npm-run-all --sequential build:clean generate:i18n build:types build:source build:testingSelectors",
     "dev:source": "yarn build:source --watch",
     "dev:types": "yarn build:types --watch",
     "dev": "npm-run-all --parallel dev:source dev:types",
